+++ nebuchadnezzar/nebu/tests/cli/test_main.py
@@ -419,3 +419,41 @@ def test_publish_cmd_with_errors(datadir, monkeypatch,
     #       the last line so we know we got to the correct place.
     # assert result.output == expected_output
     assert expected_output in result.output
+
+
+class TestConfigAtomCmd:
+
+    @pytest.fixture(autouse=True)
+    def setup(self, tmpdir, monkeypatch):
+        import pkg_resources
+        self.resource_filename = 'foo/bar.rng'
+        monkeypatch.setattr(pkg_resources, 'resource_filename',
+                            lambda *a: self.resource_filename)
+        self.tmpdir = Path(str(tmpdir))
+        monkeypatch.setattr(Path, 'home', lambda: self.tmpdir)
+
+    def check(self, result):
+        assert result.exit_code == 0
+        assert 'Wrote ' in result.output
+        assert '/.atom/config.cson' in result.output
+
+        filepath = self.tmpdir / '.atom/config.cson'
+        with filepath.open('r') as fb:
+            assert self.resource_filename in fb.read()
+
+    def test(self, invoker):
+        from nebu.cli.main import cli
+        args = ['config-atom', '--yes']
+        result = invoker(cli, args)
+        self.check(result)
+
+    def test_with_exiting_config(self, monkeypatch, invoker):
+        filepath = self.tmpdir / '.atom/config.cson'
+        filepath.parent.mkdir()
+        with filepath.open('w') as fb:
+            fb.write('baz')
+
+        from nebu.cli.main import cli
+        args = ['config-atom', '--yes']
+        result = invoker(cli, args)
+        self.check(result)